requirements/api.txt
# API layer — FastAPI, Celery async workers, Redis, HTTP client
fastapi==0.111.*
uvicorn[standard]==0.29.*
pydantic==2.*
pydantic-settings==2.*
celery==5.*
redis==5.*
httpx==0.27.*

# Database
sqlalchemy==2.0.*
asyncpg==0.29.*
alembic==1.13.*

# Auth
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
python-multipart==0.0.*
